Calculate top 5 monetized sites by timeSpent and needs some love
To calculate how much an origin "needs love", get the ratio of timeSpent
to visits. i.e. "needs love ratio" = timeSpentAtOrigin / numberOfVisits.
A small ratio indicates that, relative to how many times the user visits
the monetized site, they don't seem to spend much time there. We'd like to
inform the user if this is the case, so that the user can consider spending
more time on that site to support the creator (more time on the site = more
payment streamed if they are using a payment provider). A bit more complexity
is added to determining relative "love needed" by using the NEEDS_LOVE_MAGIC_NUMBER
and NEEDS_LOVE_MAGIC_NUMBER_MARGIN constants.

Signed-off-by: sharon-wang <sharon-wang-cpsc@outlook.com>