Hide needs love by default until monetized site visited
Hide the needs love div until a monetized site is visited. Until
a monetized site is visited for the first time since installation
of the browser extension, show the resources for a non-monetization
provider user. Hide the resources if the user is using a provider,
otherwise keep displaying the resource blocks for someone not
using a Web Monetization provider.

Signed-off-by: sharon-wang <sharon-wang-cpsc@outlook.com>