Track monetized and non-monetized time properly
Previously, we assumed that as soon as we see a payment pointer on
a website (at an origin), that the origin is always considered to
be monetized. So, once we decided that an origin is monetized, any
time spent on the site was considered monetized time spent on the
site. However, this is a flawed assumption, since sites such as
YouTube may only contain payment pointers (meta tags) on specific
videos and not others. So, some of the time spent on the site is
monetized and some is non-monetized. This PR updates how we track
time by checking for a payment pointer as we track time, to store
time spent properly.

We no longer store detailed data for every single origin -- only
monetized origins are stored. Additionally, counting monetized
visits and associated functions have been removed as they were
not being used or were redundant. The `isCurrentlyMonetized` field
has been removed from AkitaOriginData as it was a major part of
the problematic monetized time tracking. The text in the top site
detail has been updated to point out that the time is monetized
time, just to be clear.

Signed-off-by: sharon-wang <sharon-wang-cpsc@outlook.com>