Add https://*/* permission to avoid fetch error in Firefox
Firefox is unable to successfully fetch (results in
"TypeError: NetworkError when attempting to fetch resource")
without having permission to https. We fetch SPSP endpoints to
check that they are valid, so that we know the payment pointers
are legitimate. The monetization icon will not be set if we cannot
validate the payment pointer.

This permission doesn't seem to be necessary on
Chromium-based browsers.

Signed-off-by: sharon-wang <sharon-wang-cpsc@outlook.com>