Fix originData favicon path not being stored
This bug was introduced in https://github.com/esse-dev/akita/pull/141.
This PR made it so that favicon fetching did not block storing
originData. This resulted in the originData object being updated after
it had already been stored, meaning the updated originData object was
never stored.

Fixing this issue requires introducing a new function
`storeOriginFavicon`which fetches the favicon, and once/if the fetch
resolves, updates and stores the originData safely using a lock.